Update ARM kernel memory functions patch
The new version of the ARM kernel memory functions patch seems to show
some progress on optimizing memcpy -- the fast-path implementation
for smaller word-aligned sizes seems to be paying off, reducing the
memcpy profile by half in certain benchmarks. Also better prefetch
strategies for the Raspberry Pi are helping. The changes to
the file copy_template.S are getting a little complex.